Add files via upload
a practical description of density estimation based on histograms and kernel methods. Kernel density estimator of an unknown density was constructed completely nonparametric in the sense that it had no preferences and worked reasonably well for all shapes An important aim was to encourage practicing statisticians to apply these methods to data using python.